+++ snowddl/resolver/__init__.py
@@ -8,7 +8,9 @@ from .alert import AlertResolver
 from .business_role import BusinessRoleResolver
 from .clone_table import CloneTableResolver
 from .database import DatabaseResolver
-from .database_access_role import DatabaseAccessRoleResolver
+from .database_owner_role import DatabaseOwnerRoleResolver
+from .database_read_role import DatabaseReadRoleResolver
+from .database_write_role import DatabaseWriteRoleResolver
 from .dynamic_table import DynamicTableResolver
 from .event_table import EventTableResolver
 from .external_access_integration import ExternalAccessIntegrationResolver
@@ -33,7 +35,9 @@ from .row_access_policy import RowAccessPolicyResolver
 from .sequence import SequenceResolver
 from .share_access_role import ShareAccessRoleResolver
 from .schema import SchemaResolver
-from .schema_access_role import SchemaAccessRoleResolver
+from .schema_owner_role import SchemaOwnerRoleResolver
+from .schema_read_role import SchemaReadRoleResolver
+from .schema_write_role import SchemaWriteRoleResolver
 from .secret import SecretResolver
 from .stage import StageResolver
 from .stage_file import StageFileResolver
@@ -46,19 +50,25 @@ from .unique_key import UniqueKeyResolver
 from .user import UserResolver
 from .user_role import UserRoleResolver
 from .warehouse import WarehouseResolver
-from .warehouse_access_role import WarehouseAccessRoleResolver
+from .warehouse_monitor_role import WarehouseMonitorRoleResolver
+from .warehouse_usage_role import WarehouseUsageRoleResolver
 
 
 default_resolve_sequence = [
     AccountParameterResolver,
     ResourceMonitorResolver,
     WarehouseResolver,
-    WarehouseAccessRoleResolver,
+    WarehouseMonitorRoleResolver,
+    WarehouseUsageRoleResolver,
     DatabaseResolver,
     SchemaResolver,
     ShareAccessRoleResolver,
-    DatabaseAccessRoleResolver,
-    SchemaAccessRoleResolver,
+    DatabaseReadRoleResolver,
+    DatabaseWriteRoleResolver,
+    SchemaReadRoleResolver,
+    SchemaWriteRoleResolver,
+    DatabaseOwnerRoleResolver,
+    SchemaOwnerRoleResolver,
     SecretResolver,
     NetworkRuleResolver,
     ExternalAccessIntegrationResolver,
@@ -105,7 +115,8 @@ default_destroy_sequence = [
     AccountParameterResolver,
     ResourceMonitorResolver,
     WarehouseResolver,
-    WarehouseAccessRoleResolver,
+    WarehouseMonitorRoleResolver,
+    WarehouseUsageRoleResolver,
     # --
     NetworkPolicyResolver,
     AuthenticationPolicyResolver,
@@ -114,8 +125,12 @@ default_destroy_sequence = [
     DatabaseResolver,
     SchemaResolver,
     ShareAccessRoleResolver,
-    DatabaseAccessRoleResolver,
-    SchemaAccessRoleResolver,
+    DatabaseReadRoleResolver,
+    DatabaseWriteRoleResolver,
+    SchemaReadRoleResolver,
+    SchemaWriteRoleResolver,
+    DatabaseOwnerRoleResolver,
+    SchemaOwnerRoleResolver,
     # --
     OutboundShareResolver,
     TechnicalRoleResolver,