+++ snowddl/resolver/database_owner_role.py
@@ -11,9 +11,12 @@ from snowddl.blueprint import (
 from snowddl.resolver.abc_role_resolver import AbstractRoleResolver, ObjectType
 
 
-class DatabaseAccessRoleResolver(AbstractRoleResolver):
+class DatabaseOwnerRoleResolver(AbstractRoleResolver):
     def get_role_suffix(self):
         return self.config.DATABASE_ACCESS_ROLE_SUFFIX
+
+    def get_role_type(self):
+        return self.config.OWNER_ROLE_TYPE
 
     def get_blueprints(self):
         blueprints = []
@@ -24,19 +27,12 @@ class DatabaseAccessRoleResolver(AbstractRoleResolver):
             if database_permission_model.ruleset.create_database_owner_role:
                 blueprints.append(self.get_blueprint_owner_role(database_bp))
 
-            if database_permission_model.ruleset.create_database_write_role:
-                blueprints.append(self.get_blueprint_write_role(database_bp))
-
-            if database_permission_model.ruleset.create_database_read_role:
-                blueprints.append(self.get_blueprint_read_role(database_bp))
-
         return {str(bp.full_name): bp for bp in blueprints}
 
     def get_blueprint_owner_role(self, database_bp: DatabaseBlueprint):
         grants = []
         account_grants = []
         future_grants = []
-        depends_on = set()
 
         database_permission_model = self.config.get_permission_model(database_bp.permission_model)
 
@@ -83,6 +79,12 @@ class DatabaseAccessRoleResolver(AbstractRoleResolver):
         for database_name_pattern in database_bp.owner_database_read:
             grants.extend(self.build_database_role_grants(database_name_pattern, self.config.READ_ROLE_TYPE))
 
+        for schema_name_pattern in database_bp.owner_schema_write:
+            grants.extend(self.build_schema_role_grants(schema_name_pattern, self.config.WRITE_ROLE_TYPE))
+
+        for schema_name_pattern in database_bp.owner_schema_read:
+            grants.extend(self.build_schema_role_grants(schema_name_pattern, self.config.READ_ROLE_TYPE))
+
         for integration_name in database_bp.owner_integration_usage:
             grants.append(self.build_integration_grant(integration_name))
 
@@ -126,140 +128,12 @@ class DatabaseAccessRoleResolver(AbstractRoleResolver):
                     )
                 )
 
-        # Add explicit dependencies on other database roles
-        for g in grants:
-            if g.on == ObjectType.ROLE and str(g.name).endswith(self.get_role_suffix()):
-                depends_on.add(g.name)
-
         bp = RoleBlueprint(
             full_name=build_role_ident(
                 self.config.env_prefix, database_bp.full_name.database, self.config.OWNER_ROLE_TYPE, self.get_role_suffix()
             ),
             grants=grants,
             account_grants=account_grants,
-            future_grants=future_grants,
-            depends_on=depends_on,
-        )
-
-        return bp
-
-    def get_blueprint_read_role(self, database_bp: DatabaseBlueprint):
-        grants = []
-        future_grants = []
-
-        database_permission_model = self.config.get_permission_model(database_bp.permission_model)
-
-        grants.append(
-            Grant(
-                privilege="USAGE",
-                on=ObjectType.DATABASE,
-                name=database_bp.full_name,
-            )
-        )
-
-        future_grants.append(
-            FutureGrant(
-                privilege="USAGE",
-                on_future=ObjectType.SCHEMA,
-                in_parent=ObjectType.DATABASE,
-                name=database_bp.full_name,
-            )
-        )
-
-        # Future grants on DATABASE level
-        for model_future_grant in database_permission_model.read_future_grants:
-            future_grants.append(
-                FutureGrant(
-                    privilege=model_future_grant.privilege,
-                    on_future=model_future_grant.on,
-                    in_parent=ObjectType.DATABASE,
-                    name=database_bp.full_name,
-                )
-            )
-
-        # Future grants for each known schema on SCHEMA level
-        # It is required to alleviate SCHEMA level grants overriding DATABASE level
-        # https://docs.snowflake.com/en/sql-reference/sql/grant-privilege#future-grants-on-database-or-schema-objects
-        for schema_bp in self.config.get_blueprints_by_type(SchemaBlueprint).values():
-            if database_bp.full_name != schema_bp.full_name.database_full_name:
-                continue
-
-            for model_future_grant in database_permission_model.read_future_grants:
-                future_grants.append(
-                    FutureGrant(
-                        privilege=model_future_grant.privilege,
-                        on_future=model_future_grant.on,
-                        in_parent=ObjectType.SCHEMA,
-                        name=schema_bp.full_name,
-                    )
-                )
-
-        bp = RoleBlueprint(
-            full_name=build_role_ident(
-                self.config.env_prefix, database_bp.full_name.database, self.config.READ_ROLE_TYPE, self.get_role_suffix()
-            ),
-            grants=grants,
-            future_grants=future_grants,
-        )
-
-        return bp
-
-    def get_blueprint_write_role(self, database_bp: DatabaseBlueprint):
-        grants = []
-        future_grants = []
-
-        database_permission_model = self.config.get_permission_model(database_bp.permission_model)
-
-        grants.append(
-            Grant(
-                privilege="USAGE",
-                on=ObjectType.DATABASE,
-                name=database_bp.full_name,
-            )
-        )
-
-        future_grants.append(
-            FutureGrant(
-                privilege="USAGE",
-                on_future=ObjectType.SCHEMA,
-                in_parent=ObjectType.DATABASE,
-                name=database_bp.full_name,
-            )
-        )
-
-        # Future grants on DATABASE level
-        for model_future_grant in database_permission_model.write_future_grants:
-            future_grants.append(
-                FutureGrant(
-                    privilege=model_future_grant.privilege,
-                    on_future=model_future_grant.on,
-                    in_parent=ObjectType.DATABASE,
-                    name=database_bp.full_name,
-                )
-            )
-
-        # Future grants for each known schema on SCHEMA level
-        # It is required to alleviate SCHEMA level grants overriding DATABASE level
-        # https://docs.snowflake.com/en/sql-reference/sql/grant-privilege#future-grants-on-database-or-schema-objects
-        for schema_bp in self.config.get_blueprints_by_type(SchemaBlueprint).values():
-            if database_bp.full_name != schema_bp.full_name.database_full_name:
-                continue
-
-            for model_future_grant in database_permission_model.write_future_grants:
-                future_grants.append(
-                    FutureGrant(
-                        privilege=model_future_grant.privilege,
-                        on_future=model_future_grant.on,
-                        in_parent=ObjectType.SCHEMA,
-                        name=schema_bp.full_name,
-                    )
-                )
-
-        bp = RoleBlueprint(
-            full_name=build_role_ident(
-                self.config.env_prefix, database_bp.full_name.database, self.config.WRITE_ROLE_TYPE, self.get_role_suffix()
-            ),
-            grants=grants,
             future_grants=future_grants,
         )