+++ snowddl/resolver/schema_owner_role.py
@@ -10,9 +10,12 @@ from snowddl.blueprint import (
 from snowddl.resolver.abc_role_resolver import AbstractRoleResolver, ObjectType
 
 
-class SchemaAccessRoleResolver(AbstractRoleResolver):
+class SchemaOwnerRoleResolver(AbstractRoleResolver):
     def get_role_suffix(self):
         return self.config.SCHEMA_ACCESS_ROLE_SUFFIX
+
+    def get_role_type(self):
+        return self.config.OWNER_ROLE_TYPE
 
     def get_blueprints(self):
         blueprints = []
@@ -23,19 +26,12 @@ class SchemaAccessRoleResolver(AbstractRoleResolver):
             if schema_permission_model.ruleset.create_schema_owner_role:
                 blueprints.append(self.get_blueprint_owner_role(schema_bp))
 
-            if schema_permission_model.ruleset.create_schema_write_role:
-                blueprints.append(self.get_blueprint_write_role(schema_bp))
-
-            if schema_permission_model.ruleset.create_schema_read_role:
-                blueprints.append(self.get_blueprint_read_role(schema_bp))
-
         return {str(bp.full_name): bp for bp in blueprints}
 
     def get_blueprint_owner_role(self, schema_bp: SchemaBlueprint):
         grants = []
         account_grants = []
         future_grants = []
-        depends_on = set()
 
         schema_permission_model = self.config.get_permission_model(schema_bp.permission_model)
 
@@ -123,11 +119,6 @@ class SchemaAccessRoleResolver(AbstractRoleResolver):
                 )
             )
 
-        # Add explicit dependencies on other schema roles
-        for g in grants:
-            if g.on == ObjectType.ROLE and str(g.name).endswith(self.get_role_suffix()):
-                depends_on.add(g.name)
-
         bp = RoleBlueprint(
             full_name=build_role_ident(
                 self.config.env_prefix,
@@ -138,99 +129,6 @@ class SchemaAccessRoleResolver(AbstractRoleResolver):
             ),
             grants=grants,
             account_grants=account_grants,
-            future_grants=future_grants,
-            depends_on=depends_on,
-        )
-
-        return bp
-
-    def get_blueprint_read_role(self, schema_bp: SchemaBlueprint):
-        grants = []
-        future_grants = []
-
-        schema_permission_model = self.config.get_permission_model(schema_bp.permission_model)
-
-        grants.append(
-            Grant(
-                privilege="USAGE",
-                on=ObjectType.DATABASE,
-                name=DatabaseIdent(schema_bp.full_name.env_prefix, schema_bp.full_name.database),
-            )
-        )
-
-        grants.append(
-            Grant(
-                privilege="USAGE",
-                on=ObjectType.SCHEMA,
-                name=schema_bp.full_name,
-            )
-        )
-
-        for model_future_grant in schema_permission_model.read_future_grants:
-            future_grants.append(
-                FutureGrant(
-                    privilege=model_future_grant.privilege,
-                    on_future=model_future_grant.on,
-                    in_parent=ObjectType.SCHEMA,
-                    name=schema_bp.full_name,
-                )
-            )
-
-        bp = RoleBlueprint(
-            full_name=build_role_ident(
-                self.config.env_prefix,
-                schema_bp.full_name.database,
-                schema_bp.full_name.schema,
-                self.config.READ_ROLE_TYPE,
-                self.get_role_suffix(),
-            ),
-            grants=grants,
-            future_grants=future_grants,
-        )
-
-        return bp
-
-    def get_blueprint_write_role(self, schema_bp: SchemaBlueprint):
-        grants = []
-        future_grants = []
-
-        schema_permission_model = self.config.get_permission_model(schema_bp.permission_model)
-
-        grants.append(
-            Grant(
-                privilege="USAGE",
-                on=ObjectType.DATABASE,
-                name=DatabaseIdent(schema_bp.full_name.env_prefix, schema_bp.full_name.database),
-            )
-        )
-
-        grants.append(
-            Grant(
-                privilege="USAGE",
-                on=ObjectType.SCHEMA,
-                name=schema_bp.full_name,
-            )
-        )
-
-        for model_future_grant in schema_permission_model.write_future_grants:
-            future_grants.append(
-                FutureGrant(
-                    privilege=model_future_grant.privilege,
-                    on_future=model_future_grant.on,
-                    in_parent=ObjectType.SCHEMA,
-                    name=schema_bp.full_name,
-                )
-            )
-
-        bp = RoleBlueprint(
-            full_name=build_role_ident(
-                self.config.env_prefix,
-                schema_bp.full_name.database,
-                schema_bp.full_name.schema,
-                self.config.WRITE_ROLE_TYPE,
-                self.get_role_suffix(),
-            ),
-            grants=grants,
             future_grants=future_grants,
         )